requirements.txt
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
requests==2.32.3
beautifulsoup4==4.12.3
google==3.0.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pytz==2024.1
langchain==0.2.12
langchain-chroma==0.1.2
langchain-community==0.2.11
langchain-groq==0.1.9
langchain-google-genai==1.0.8
langchain-huggingface==0.0.3
streamlit==1.37.1
schedule==1.2.2
lxml-html-clean
newspaper3k